class breakdown first pass
