+++ #README.md#
@@ -1,0 +1,12 @@
+This is a tool to aid in the creation of a paper concerning the computational complexity of the game Popils for the Sega Game Gear.
+
+TODOs
+- Clean up/factorize main game loop
+- Split main file into multiple smaller files
+- Refactor argument parsing into functions
+- Make Player class with methods and attributes
+- Make AI class for different solvers
+- Scroll screen for bigger puzzles
+- Print prettier and more descriptive output (CNF 3SAT, solution, etc.)
+- Label blocks in puzzle for variables
+- Create more example files
