+++ block.py
@@ -1,0 +1,3 @@
+class Block():
+    def __init__(self, color, connectivity=[], destructible=False):
+        pass
