+++ game.py
@@ -1,0 +1,2 @@
+class Game: #tile-based game, either popils or megalit
+    pass
